squash a -Wfloat-conversion warning

+++ cmd/smyrna/glmotion.c
@@ -30,7 +30,7 @@ static void graph_zoom(float real_zoom)
 		old_zoom = view->cameras[view->active_camera]->r;
 
     if (real_zoom < view->Topview->fitin_zoom * MAX_ZOOM)
-		real_zoom = (float) view->Topview->fitin_zoom * MAX_ZOOM;
+		real_zoom = view->Topview->fitin_zoom * MAX_ZOOM;
     if (real_zoom > view->Topview->fitin_zoom * MIN_ZOOM)
 		real_zoom = view->Topview->fitin_zoom * MIN_ZOOM;
     if (view->active_camera == -1)
